requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.0.4
django-cors-headers==4.7.0
django-import-export==4.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fonttools==4.56.0
fpdf2==2.8.2
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
psutil==7.0.0
PyJWT==2.9.0
PyPDF2==3.0.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tablib==3.8.0
uritemplate==4.1.1
urllib3==2.3.0
celery==5.4.0
redis==5.2.1
django-celery-beat==2.6.0